requirements.txt
# Core encryption
cryptography>=41.0.0
python-dotenv>=1.0.0
watchdog>=3.0.0

# CLI
typer[all]>=0.9.0
rich>=13.0.0
questionary>=2.0.1

# Cloud Providers
## Google Drive
google-api-python-client>=2.120.0
google-auth-httplib2
google-auth-oauthlib
## Backblaze B2
b2sdk>=1.20.0

# Utilis
tqdm>=4.66.0
requests>=2.31.0

# Tests
pytest>=8.0.0